Added HashedLocator
Added struct to quickly locate uv given surf and x. A hash array is filled with uv values given from `x=lower` to `x=upper` with resolution `step`. The hash is linearly interpolated at x_new, and this uv is returned if x_new is outside the hash domain. Otherwise, the uv guess is improved with a AD Newton step.

Still needs a method to update the hash as the surf evolves in time.