Added map to ParametricBodies
You can supply a map from (x,t) to the surface coordinate ξ. This lets you define rigid body motion, scalings, reflections, tilings, etc... and apply them to parametrically defined bodies. The body can still be moving/deforming, but these need to be defined in terms of ξ.

One limitation I've built-in is that the scale needs to be a constant since I've precomputed this scale when defining the body. So you can can to the desired resolution and that's it.