Clamp newton refinement
Newton's method is unstable when two roots are close together. Now that we're a little less sensitive to uv errors, we can clamp the step amplification to get a "good-enough" uv with much better numerical stability. Has to lower the test tolerances a tiny bit, but everything is still plenty accurate.