Fixed (?) type stability
I think I have fixed the type stability of the `bspline` function. However, it does require that the element type of the control point data be "less than" or matching the type of the parameter (s). I could probably automate recasting the data when making a ParametricBody, but I am not sure about the best approach.

Surprisingly, the accuracy of the `measure` function is really bad. I have not looked into why yet.