Further progress
I fixed another bug in the setup (curves need to wind CCW), and it's now running correctly on the CPU. It doesn't compile on the GPU and the error isn't helpful. It is dying when trying to call the `align` function within `refine`.

One issue is that HashedLocator requires that `curve` returns an SA. The CPU is nimble enough to deal with this gracefully, but the GPU isn't. So I added this requirement back into the initialization, but it's not enough to solve the problem.

I think the more fundamental issue is that the control point array is not on the GPU. I'll try to adapt the function to the GPU before passing it into HashedLocator.