Bug fix for closed splines
Closed splines need to use a mod parameter instead of being clamped. I changed `refine` to use the correct type of arithmetic depending on the geometry, but I don't see a simple way to interpolate across the modular boundary. Instead, I've switched to using the nearest hash point as the initial estimate.