Update toml and HashedLocator
I removed the interpolation functionality and regained the lost accuracy with a second Newton step. This is _much_ more expensive than interpolation, but much more robust to "rough" hash values and should also be more precise. May be worth revisting later for additional speed-up.

Also the package was trying to use CUDA 5.0 and an old version of WaterLily, so I've forced WaterLily version [1,2) in the toml.