Got Bsplines running on GPUs
Parameterizing the spline type by the degree and then using that as a static parameter in the recursive Cox de Boor allows the GPU to compile and run! Almost all tests passing.

The derivatives at the end points seem to be broken, and this is causing the location test to fail.