Fix derivatives
I had to move the u checks back into the B{0} function to get the derivative working. This is less efficient since you will calculate a lot of B{>0} values which ultimately get multiplied by B{0}=0. I limited this a bit by adding checks to the top level loop. It's still not optimal, but I guess those extra computations are used in calculating the derivatives.