updated normal calculation
I decided to get rid of the dis function and always use p'*n-thk, adding thk as a struct parameter. This always works if the normal is pointing into the flow, which requires one extra parameter `signed`, to help decide when to use `norm_dir` and when to use the new `aligned_dir` function.