Specialize the NurbsLocator
The nurbs locator broke hard for degree=1 curves because d^2 is a very strong function of u, so checking a few points on a segment is not a good way to check if the segment is the closest. Replaced with a specialized degree=1 locator using inverse quadratic interpolation on each segment (leading to the exact minimizer) which can then be compared robustly.

The same issue can happen on degree>1 splines (just less commonly). New version quick checks a segment (using a single Davidon step) and continues the Davidon search if the new segment has d^2 near the current best. This makes it much less likely you miss the right segment.

Passing all tests (back to high precision) and gives perfect visual SDFs on whale tale test for degree=1:4.